+++ Tasks/UnityCMD/UnityCMDV1/unity-cmd.ts
@@ -3,8 +3,8 @@ import tl = require('azure-pipelines-task-lib/task');
 import {
     UnityToolRunner,
     UnityPathTools,
-    UnityLogTools
-} from '@dinomite-studios/unity-utilities';
+    Utilities
+} from '@dinomite-studios/unity-azure-pipelines-tasks-lib';
 import { getUnityEditorVersion } from './unity-build-shared';
 
 tl.setResourcePath(path.join(__dirname, 'task.json'));
@@ -30,10 +30,10 @@ async function run() {
             tl.getInput(unityEditorsPathModeInputVariableName, true)!,
             tl.getInput(customUnityEditorsPathInputVariableName));
         const unityVersion = getUnityEditorVersion();
-        const unityExecutablePath = UnityPathTools.getUnityExecutableFullPath(unityEditorsPath, unityVersion);
+        const unityExecutablePath = UnityPathTools.getUnityExecutableFullPath(unityEditorsPath, unityVersion.info!);
         const repositoryLocalPath = tl.getVariable(localPathInputVariableName)!;
         const logFilesDirectory = path.join(repositoryLocalPath!, 'Logs');
-        const logFilePath = path.join(logFilesDirectory, `UnityCMDLog_${UnityLogTools.getLogFileNameTimeStamp()}.log`);
+        const logFilePath = path.join(logFilesDirectory, `UnityCMDLog_${Utilities.getLogFileNameTimeStamp()}.log`);
 
         // Set output variable values.
         tl.setVariable(logsOutputPathOutputVariableName, logFilesDirectory);