+++ Tasks/UnityGetProjectVersion/UnityGetProjectVersionV1/unity-get-project-version.ts
@@ -1,6 +1,6 @@
 import path = require('path');
 import tl = require('azure-pipelines-task-lib/task');
-import { ProjectVersionService } from '@dinomite-studios/unity-project-version';
+import { UnityVersionTools } from '@dinomite-studios/unity-azure-pipelines-tasks-lib';
 
 tl.setResourcePath(path.join(__dirname, 'task.json'));
 
@@ -21,7 +21,7 @@ function run() {
         console.log(`${tl.loc('projectPathInfo')} ${projectPath}`);
 
         // Determine the project's last used Unity editor version.
-        const unityVersion = ProjectVersionService.determineProjectVersionFromFile(projectPath);
+        const unityVersion = UnityVersionTools.determineProjectVersionFromFile(projectPath);
         if (unityVersion.error) {
             const error = `${tl.loc('failGetUnityEditorVersion')} | ${unityVersion.error}`;
             console.error(error);
@@ -29,16 +29,16 @@ function run() {
         }
 
         // Log findings.
-        const successGetVersionLog = `${tl.loc('successGetUnityEditorVersion')} ${unityVersion.version}${unityVersion.revision ? `, revision=${unityVersion.revision}` : ''}, alpha=${unityVersion.isAlpha}, beta=${unityVersion.isBeta}`;
+        const successGetVersionLog = `${tl.loc('successGetUnityEditorVersion')} ${unityVersion.info!.version}${unityVersion.info!.revision ? `, revision=${unityVersion.info!.revision}` : ''}, alpha=${unityVersion.info!.isAlpha}, beta=${unityVersion.info!.isBeta}`;
         console.log(successGetVersionLog);
-        if (unityVersion.isAlpha || unityVersion.isBeta) {
+        if (unityVersion.info!.isAlpha || unityVersion.info!.isBeta) {
             console.warn(tl.loc('warningAlphaBetaVersion'));
         }
 
         // Set output variable values.
-        tl.setVariable(projectVersionOutputVariableName, unityVersion.version);
-        if (unityVersion.revision) {
-            tl.setVariable(projectVersionRevisionOutputVariableName, unityVersion.revision);
+        tl.setVariable(projectVersionOutputVariableName, unityVersion.info!.version);
+        if (unityVersion.info!.revision) {
+            tl.setVariable(projectVersionRevisionOutputVariableName, unityVersion.info!.revision);
         }
 
         // Set task result succeeded.