+++ Tasks/UnityTest/UnityTestV1/unity-build-shared.ts
@@ -1,25 +1,23 @@
 import tl = require('azure-pipelines-task-lib/task');
-import {
-    UnityVersionTools,
-    UnityVersionInfo
-} from '@dinomite-studios/unity-azure-pipelines-tasks-lib';
+import { UnityVersionInfoResult, UnityVersionTools } from '@dinomite-studios/unity-azure-pipelines-tasks-lib/';
 
-export function getUnityEditorVersion(): UnityVersionInfo {
+export function getUnityEditorVersion(): UnityVersionInfoResult {
     const projectPath = tl.getPathInput('unityProjectPath') || '';
     console.log(`${tl.loc('projectPathInfo')} ${projectPath}`);
 
-    const unityVersionResult = UnityVersionTools.determineProjectVersionFromFile(projectPath);
-    if (unityVersionResult.error || !unityVersionResult.info) {
-        const error = `${tl.loc('failGetUnityEditorVersion')} | ${unityVersionResult.error}`;
+    const unityVersion = UnityVersionTools.determineProjectVersionFromFile(projectPath);
+    if (unityVersion.error) {
+        const error = `${tl.loc('failGetUnityEditorVersion')} | ${unityVersion.error}`;
         console.error(error);
         throw new Error(error);
     }
 
-    const successGetVersionLog = `${tl.loc('successGetUnityEditorVersion')} ${unityVersionResult.info!.version}${unityVersionResult.info!.revision ? `, revision=${unityVersionResult.info!.revision}` : ''}, alpha=${unityVersionResult.info!.isAlpha}, beta=${unityVersionResult.info!.isBeta}`;
+    const successGetVersionLog = `${tl.loc('successGetUnityEditorVersion')} ${unityVersion.info!.version}${unityVersion.info!.revision ? `, revision=${unityVersion.info!.revision}` : ''}, alpha=${unityVersion.info!.isAlpha}, beta=${unityVersion.info!.isBeta}`;
     console.log(successGetVersionLog);
-    if (unityVersionResult.info!.isAlpha || unityVersionResult.info!.isBeta) {
+
+    if (unityVersion.info!.isAlpha || unityVersion.info!.isBeta) {
         console.warn(tl.loc('warningAlphaBetaVersion'));
     }
 
-    return unityVersionResult.info!;
+    return unityVersion;
 }