+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.9.0]
+
+### Changed
+
+- Upgrade all tasks to Node 10
+- Updated minimum required agent version to `2.144.0`
+- Updated UnityTest Task to v1.5.0 [Changelog](https://github.com/Dinomite-Studios/unity-azure-pipelines-tasks/blob/main/Tasks/UnityTest/UnityTestV1/CHANGELOG.md)
+- Updated UnityGetProjectVersion Task to v1.1.0 [Changelog](https://github.com/Dinomite-Studios/unity-azure-pipelines-tasks/blob/main/Tasks/UnityGetProjectVersion/UnityGetProjectVersionV1/CHANGELOG.md)
+- Updated UnityCMD Task to v1.2.0 [Changelog](https://github.com/Dinomite-Studios/unity-azure-pipelines-tasks/blob/main/Tasks/UnityCMD/UnityCMDV1/CHANGELOG.md)
+- Updated UnityActivateLicense Task to v1.1.0 [Changelog](https://github.com/Dinomite-Studios/unity-azure-pipelines-tasks/blob/main/Tasks/UnityActivateLicense/UnityActivateLicenseV1/CHANGELOG.md)
+- Updated UnityBuild Task to v3.2.0 [Changelog](https://github.com/Dinomite-Studios/unity-azure-pipelines-tasks/blob/main/Tasks/UnityBuild/UnityBuildV3/CHANGELOG.md)
+
+### Removed
+
+- Legacy Unity Build Task V1 is no longer shipped with the extension. Please use Unity Build V3
+- Legacy Unity Build Task V2 is no longer shipped with the extension. Please use Unity Build V3
+
 ## [2.8.6]
 
 ### Changed