+++ Tasks/UnityGetProjectVersion/UnityGetProjectVersionV1/CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0]
+
+### Fixed
+
+- This task uses Node 6 execution handler, which will be deprecated soon [GitHub Issue](https://github.com/Dinomite-Studios/unity-azure-pipelines-tasks/issues/190)
+
+### Changed
+
+- Upgrade task to Node 10
+- Updated minimum required agent version to `2.144.0`
+- Use new shared library `@dinomite-studios/unity-azure-pipelines-tasks-lib` for common task features
+
 ## [1.0.32]
 
 Rerelease to fix broken build from previous release.